72 Remove in progress cleanup from compression procedure
The current implementation is insufficient and error-prone, as there
is no information which node actually is progressing the txs, so in
case there are e.g. two nodes one lost and one dead with selected
to commit the tx and died after that, the existing approach would
not clean the tx from the in progress set.