78 Rework KafkaFactoryForTests to ensure topics existence for consumer
After investigation, it appeared that requests from consumer does not
lead to topic creation by broker, though it is not mentioned
in the documentation of property controlling auto creation possibility.

Also if the first poll is made by consumer to such an unexistent topics,
subsequent polls seems to return nothing even after the topic was
created, e.g. by publishing messages to it.

So the approach was taken to wrap a consumer and upon subscription
access each topic with in-time created producer to trigger their
auto-creation.